refactor(stereo): fold EuRoC rectifier adapter into euroc.rs, drop rectify.rs
The `rectify.rs` module was down to just the EuRoC `T_BS` adapter plus
re-exports of the upstream rectifier. Move `rectifier_from_euroc` (and
its private helpers + test) into `datasets/euroc.rs` next to
`EurocCameraCalibration`, re-export `StereoRectifier` straight from
`kornia_3d::stereo`, and delete the file.
