+++ examples/common/datasets/mod.rs
@@ -1,7 +1,9 @@
 pub mod euroc;
-pub mod rectify;
 pub mod stereo_calib;
 
-pub use euroc::EurocDataset;
-pub use rectify::{StereoRectifier, rectifier_from_euroc};
+// The Bouguet stereo rectifier lives upstream in kornia-3d; re-export it here so
+// dataset code can refer to `crate::datasets::StereoRectifier`. The EuRoC `T_BS`
+// adapter that feeds it lives with the EuRoC reader in `euroc`.
+pub use euroc::{EurocDataset, rectifier_from_euroc};
+pub use kornia_3d::stereo::StereoRectifier;
 pub use stereo_calib::StereoCalib;
