+++ examples/common/datasets/stereo_calib.rs
@@ -15,7 +15,7 @@ use kornia_algebra::{Mat3F64, Vec3F64};
 use kornia_imgproc::calibration::distortion::PolynomialDistortion;
 use serde::Deserialize;
 
-use super::rectify::{CameraCalib, StereoError, StereoRectifier};
+use kornia_3d::stereo::{CameraCalib, StereoError, StereoRectifier};
 
 #[derive(Deserialize)]
 struct CamYaml {
